### 2025-03-03 #7868 @warrickball
Hi @sarahstangl-lanl, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

For the record, this was first submitted in #6279 and rejected in a scope query, then withdrawn in #6637. It looks like a sticking point in the past was installation but I confirm that I was just able to install the software and start the example in the README.md on my local cluster, which I used because it was easier to get the required libraries. The first time I ran `./nudustc++` nothing appeared to happened but a second attempt worked correctly. This is certainly good enough to warrant a review but I expect this issue will come up. (Note also that the example says to run `./nudust++` rather than `./nudustc++`.)

Note also that some references in the paper are missing journal names. This is usually because the BibTeX has been copied from NASA ADS, which uses shorthands (e.g. `\apj` for ApJ). JOSS doesn't recognise these and anyway our style is to write journal names in full.
