### 2025-03-10 76fc8c0
Update paper.bib
fixed journal shortcuts in paper.bib
