+++ paper.bib
@@ -16,7 +16,7 @@ archivePrefix = {ascl},
 @ARTICLE{Draine2003,
        author = {{Draine}, B.~T.},
         title = "{Interstellar Dust Grains}",
-      journal = {\araa},
+      journal = {Annual Review of Astronomy & Astrophysics},
      keywords = {Astrophysics},
          year = 2003,
         month = jan,
@@ -44,7 +44,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Bianchi2007,
        author = {{Bianchi}, Simone and {Schneider}, Raffaella},
         title = "{Dust formation and survival in supernova ejecta}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {shock waves, supernovae: individual: Cassiopeia A, dust, extinction, supernova remnants, Astrophysics},
          year = 2007,
         month = jul,
@@ -62,7 +62,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Todini2001,
        author = {{Todini}, Paolo and {Ferrara}, Andrea},
         title = "{Dust formation in primordial Type II supernovae}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {STARS: FORMATION, SUPERNOVAE: GENERAL, SUPERNOVAE: INDIVIDUAL: SN1987A, DUST, EXTINCTION, SUPERNOVA REMNANTS, Astrophysics},
          year = 2001,
         month = aug,
@@ -135,7 +135,7 @@ archivePrefix = {arXiv},
 @ARTICLE{brooker2021,
        author = {{Brooker}, Ezra S. and {Stangl}, Sarah M. and {Mauney}, Christopher M. and {Fryer}, C.~L.},
         title = "{Dependence of Dust Formation on the Supernova Explosion}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Astrophysical dust processes, Core-collapse supernovae, 99, 304, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2022,
         month = jun,
@@ -160,7 +160,7 @@ archivePrefix = {arXiv},
 @ARTICLE{dustpy,
        author = {{Stammler}, Sebastian M. and {Birnstiel}, Tilman},
         title = "{DustPy: A Python Package for Dust Evolution in Protoplanetary Disks}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Protoplanetary disks, Astronomy software, Astronomical simulations, Circumstellar dust, Planet formation, Planetesimals, 1300, 1855, 1857, 236, 1241, 1259, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2022,
         month = aug,
