### 2025-03-10 df8e98f
Update paper.bib
fixed missing doi
