+++ paper.bib
@@ -37,6 +37,7 @@ archivePrefix = {arXiv},
   volume={77},
   number={6},
   pages={1402--1410},
+  doi = {10.1143/PTP.77.1402},
   year={1987},
   publisher={Oxford University Press}
 }
@@ -77,17 +78,19 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@INPROCEEDINGS{paquette2023,
-       author = {{Paquette}, John and {Nuth}, Joseph and {Ferguson}, Frank},
+@ARTICLE{2023ApJ...957...35P,
+       author = {{Paquette}, John A. and {Nuth}, Joseph A. and {Ferguson}, Frank T.},
         title = "{Dust Coagulation in Oxygen-rich Circumstellar Outflows}",
-    booktitle = {American Astronomical Society Meeting Abstracts},
+      journal = {The Astrophysical Journal},
+     keywords = {Circumstellar dust, Stellar winds, 236, 1636},
          year = 2023,
-       series = {American Astronomical Society Meeting Abstracts},
-       volume = {55},
-        month = sep,
-          eid = {234.04},
-        pages = {234.04},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023AAS...24223404P},
+        month = nov,
+       volume = {957},
+       number = {1},
+          eid = {35},
+        pages = {35},
+          doi = {10.3847/1538-4357/acf5e3},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...957...35P},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
