### 2025-04-15 #7868 @sarahstangl-lanl
Hello,

Have any editors become available? Is this still waitlisted?

Also, I fixed the errors in the bib file and README. I'm not sure what happened with the executable starting the second time. I'm unable to recreate this but will try on other machines.
