### 2025-04-28 #7868 @warrickball
Hi @sarahstangl-lanl, sorry for the long wait. The Astronomy, Astrophysics & Space Science (AASS) track of JOSS has a substantial backlog, which is typically leading to a ~3 month wait before an editor becomes available.

I confirm that your submission is indeed in the waiting list. The note I made above about the previous scope query is mainly context for whichever editor does eventually pick this up.

I've personally been away for most of April, so I apologise for taking longer than usual to reply.
