### 2025-06-05 #8349 @editorialbot
[REVIEW]: nuDustC: C++ Code for Modeling Dust Nucleation and Destruction in Gaseous Systems
**Submitting author:** @sarahstangl-lanl (<a href="http://orcid.org/0000-0001-5570-6666">Sarah Stangl</a>)
**Repository:** https://github.com/lanl/nuDustC
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
