### 2025-06-05 #8349 @steven-murray
I just re-installed `nudustc` and the experience was *far* better than last time -- great job!

I have a few recommendations, mostly surrounding documentation and the paper itself:

- [ ] There is a line in the paper: "but is increasingly less so dus to these limitations. (?) grain a" I expect "dus" is supposed to be "due" and then maybe there is a missing ref?
- [ ] Input description
- [ ] Typo in "for each cell with not data flow between cells" (not --> no)
- [ ] The first word of the "Performance and Accuracy" section is not capitalized (the)
- [ ] The section on "Performance and Accuracy" does not have any info about accuracy in it -- I suggest removing the "Accuracy" part of the title, or, better, adding a plot showing how accurate the solver is (e.g. compared to snDust)
- [ ] I was a bit confused by the performance figure -- the legend labels do not make it very obvious which line is nudustc vs. which are from snDust.
- [ ] The performance plot is a function of "Solution Vector Length", but it was not clear to me what this parameter is, and neither do I see a parameter in the configuration ini file that obviously maps to this. It would be great if this could be elaborated a little more (otherwise I cannot reproduce the performance claims here).
- [ ] Following from my above point, the documentation in the README file has improved considerably, however I found that the description of some of the parameters was a little terse (perhaps this reflects my perspective from outside the field, but I think it would be helpful to be a little more descriptive about some of the parameters). There are also some typos and spelling errors in the README.
- [ ] There is no description of the output file format in the README. I suggest adding one.
- [ ] I think it would be very helpful to have a "real world" tutorial. While having a "test" config.ini is great for ensuring that the code built correctly, it is not so helpful in understanding *what* the code is doing. This may be painfully obvious to someone in the field, but nevertheless I think it's good to have a little guided tutorial, discussing each of the inputs and why you would set it to a particular value, and what its effect on the results are. I understand this is a bit more work, so I'm happy to be argued out of this one ;-).
- [ ] In terms of the installation, I found that everything proceeded smoothly *except* that I had to update `CMakeLists.txt` to remove the pinned version of PLOG (I think it was pinned to 1.1.9). That version has an old and unsupported version specification for cmake. Unpinning it worked for me, though perhaps pinning to a more recent version would be better.
