### 2025-06-05 9ee9209
Update paper.md
a couple spelling mistakes fixed
