+++ paper.md
@@ -48,7 +48,7 @@ This project originated from the need to track dust nucleation and destruction i
 
 In order to gain a deeper understanding of the origin and characteristics of dust in the interstellar medium, it is imperative to develop models that include the nucleation, chemistry, growth, and erosion of dust. Current methods of calculating dust formation and survival include Classical Nucleation Theory (CNT) and Kinetic Nucleation Theory (KNT).
 
-CNT treats grain formation as a barrier crossing problem. As atoms stick to a cluster, the free energy increases. After reaching a critical size, the free energy decreases as atoms are added. It tracks the nucleation rate by assuming a steady state between monomer attachment and detachment. However, it neglects chemical reactions of formation, destruction, growth by coagulation, and treats the grains as bulk materials. Due to these simplified assumptions, CNT is widely used, but is increasingly less so dus to these limitations.  @kozasa1987 grain and @Bianchi2007 used CNT to model dust grain formation in SN 1987A and SN 1987A-like Supernovae. @Todini2001 used CNT to calculate dust formation in Core Collapse Supernovae Explosions. More recently @paquette2023 used CNT to model dust formation in the outflows of AGB Stars. 
+CNT treats grain formation as a barrier crossing problem. As atoms stick to a cluster, the free energy increases. After reaching a critical size, the free energy decreases as atoms are added. It tracks the nucleation rate by assuming a steady state between monomer attachment and detachment. However, it neglects chemical reactions of formation, destruction, growth by coagulation, and treats the grains as bulk materials. Due to these simplified assumptions, CNT is widely used, but is increasingly less so due to these limitations.  @kozasa1987 grain and @Bianchi2007 used CNT to model dust grain formation in SN 1987A and SN 1987A-like Supernovae. @Todini2001 used CNT to calculate dust formation in Core Collapse Supernovae Explosions. More recently @paquette2023 used CNT to model dust formation in the outflows of AGB Stars. 
 
 KNT tracks the number densities of clusters with more than two atoms, treating them as particles that grow as spheres. It uses size dependent grain properties in place of bulk material properties. Grains grow by accreting atoms (condensation) through kinetic theory. Grains lose atoms through erosion (destruction) using the principle of detailed balance. While KNT does not assume a steady state between condensation and destruction, it still doesn't take into account the chemical reactions undergone or growth through coagulation. @Nozawa_2003 used KNT to model the nucleation and growth of dust in early Population III star supernovae. @Nozawa2006 included destruction to model the effects of high velocity shock waves on dust. @fallest2012 predicts dust mass yields in CCSNe using KNT.
 
@@ -62,9 +62,9 @@ Because of the post processing nature of <span style="font-family:qcr;">nudustc+
 
 Because <span style="font-family:qcr;">nudustc++</span> is a post processing code where data is read in separately for each cell with not data flow between cells, there is no shared memory and each cell can be computed independently. This results in an embarrassingly parallel process, allowing for simultaneous computation of each cell and thereby reducing runtime. 
 
-# Performance and Accuracy
+# Performance
 
-the Figure below shows a comparison between the python <span style="font-family:qcr;">snDust</span> using an implicit integrator versus <span style="font-family:qcr;">nudustc++</span> using an explicit integrator. At shorter solution lengths, the implicit integrator has a lot of overhead leading to an increased run time despite the shorter length. The explicit integrator in <span style="font-family:qcr;">C++</span> out performs the implicit integrator even at longer solution length. Overall, <span style="font-family:qcr;">nudustc++</span> outperforms the python version. It is also embarrassingly parallel. Utilizing OpenMP and MPI, the user can run many cells in parallel. 
[…]
