### 2025-06-05 59516d0
Update paper.md
a couple more spelling mistakes fixed
