### 2025-06-05 0cc3c98
Update paper.md
grammar and spelling fixes
