+++ paper.md
@@ -40,7 +40,7 @@ We introduce <span style="font-family:qcr;">nudustc++</span>, a **nu**cleating *
 
 Understanding interstellar dust is crucial for astronomical observations [@Draine2003], offering key insights into stellar processes. These grains absorb electromagnetic radiation, re-emitting it at longer wavelengths, leading to extinction and a spectral shift towards redder wavelengths. The size and composition of dust introduces variability in opacities and distortion of incident light, resulting in molecular lines and altering the resulting data. Dust formed in asymptotic giant branch (AGB) stars, on pre-existing grains in the interstellar medium (ISM), and within the expanding, cooling ejecta of core collapse supernovae explosions (CCSNe); these grains preserve important information about the nucleosynthetic processes within their host environment, locking up their unique isotopic signatures. However, despite their importance, the quantity, composition, and size distribution of dust formed in supernovae and deposited into the ISM remain poorly constrained. 
 
-Models of formation, growth, and weathering of dust are necessary to understand the origin and characteristics of dust in the interstellar medium, shedding light on where and what dust are possible in these environments. Specifically, modelling the formation and survivability of dust in CCSNe produce a population of dust grains that can be compared with observations, allowing verification of our current understanding of physical models: ISM dust origin, chemical networks, nucleation models, hydrodynamics, supernovae engines, progenitor structure, erosion physics, and stellar compositions. 
+Models of formation, growth, and weathering of dust are necessary to understand the origin and characteristics of dust in the interstellar medium, shedding light on where and what dust are possible in these environments. Specifically, modeling the formation and survivability of dust in CCSNe produce a population of dust grains that can be compared with observations, allowing verification of our current understanding of physical models: ISM dust origin, chemical networks, nucleation models, hydrodynamics, supernovae engines, progenitor structure, erosion physics, and stellar compositions. 
 
 This project originated from the need to track dust nucleation and destruction in Core-Collapse Supernovae Explosions. It addresses the lack of sub-grid physics associated with phase transitions, where hydrodynamical code's timesteps are an order of magnitude larger than those needed to capture gas vapor physics. A smaller more refined grid with chemical networks and smaller timesteps is needed. The code is structured to intake any hydro-dynamical temperature-density trajectory with vapor compositions. This allows <span style="font-family:qcr;">nudustc++</span> to track dust in a large range of environments: planetary atmospheres, nebulae, hydro-aerosoal formation, explosions, etc. Additionally, if a time series hydro-dynamical profile is unavailable but a dust size distribution and a profile snapshot is, <span style="font-family:qcr;">nudustc++</span> can calculate the evolution and survivability of the dust. The applications of <span style="font-family:qcr;">nudustc++</span> extend beyond Supernovae and Astronomy to include any model with thermodynamic and statistical physics. 
 
@@ -48,7 +48,7 @@ This project originated from the need to track dust nucleation and destruction i
 
 In order to gain a deeper understanding of the origin and characteristics of dust in the interstellar medium, it is imperative to develop models that include the nucleation, chemistry, growth, and erosion of dust. Current methods of calculating dust formation and survival include Classical Nucleation Theory (CNT) and Kinetic Nucleation Theory (KNT).
 
[…]
