### 2025-06-05 f19ccd3
Update README.md
spelling mistakes fixed
