### 2025-07-11 #8349 @plaplant
@sarahstangl-lanl @guadabsb15 thanks very much for your initial looks at the code! If you see any additional issues that should be addressed, please feel free to make note of them here or as issues on the [upstream repository](https://github.com/lanl/nuDustC).

@sarahstangl-lanl when you get a chance, please start addressing the issues noted above. If you have any questions, please feel free to reach out. Thanks!
