### 2025-07-15 7a70a9d
Update README.md
minor grammar and spelling updates
