+++ README.md
@@ -74,7 +74,7 @@ LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:~/miniconda3/lib ./nudustc++ -c data/inputs/t
 
 Or specifying the paths to the packages and libraries in *LD_LIBRARY_PATH*.
 
-The default had OpenMP (NUDUSTC_ENABLE_OPENMP) and MPI (NUDUSTC_ENABLE_MPI) are turned on and sundials (NUDUSTC_USE_SUNDIALS) are turned off. This can be edited in the CMakeLists.txt file. 
+OpenMP (NUDUSTC_ENABLE_OPENMP), MPI (NUDUSTC_ENABLE_MPI), and sundials (NUDUSTC_USE_SUNDIALS) are turned off. This can be edited in the CMakeLists.txt file. 
 
 If using MPI, run with
 
@@ -87,7 +87,7 @@ where 'n' is the number of processors.
 ### Inputs
 Required: Config file. This lists the various input information such as data files, integration parameters, and calculation options.
 
-# Configuration file
+# Configuration File
 
 ### Integration Parameters
 *ode_dt_0*: Initial timestep for integration.
@@ -102,13 +102,13 @@ Required: Config file. This lists the various input information such as data fil
 
     
 ### Data Files
-*sizeDist_file*: This describes the size distribution for the model. Each cell has one line. Each line is an array of size distributions of grain species in the order specified in the header line.
+*sizeDist_file*: This describes the size distribution for the model. Each cell is described in one line. Each line is an array of size distributions of grain species in the order specified in the header line.
 
-*environment_file*: This contains the trajectory data for each timestep. The time is specified on a single line. Below, each cell's is described in a single line: cell_ID, temperature (K), volume(cm^3), density(g/cm^3), pressure (Ba), velocity (cm/s), radius (cm).
+*environment_file*: This contains the trajectory data for each timestep. The time is specified on a single line. Below, each cell is described in a single line: cell_ID, temperature (K), volume(cm^3), density(g/cm^3), pressure (Ba), velocity (cm/s), radius (cm).
 
-*network_file*: This includes the chemical network of grain reactions. Each grain species takes up one line in order: reactants, "->", products, "|", key species, Gibb's free energy 'A' term (A/10^4 K), Gibb's free energy 'B' term, surface energy of the condensate (ergs/cm^2), radius of condensate (angstroms). 
+*network_file*: This includes the chemical network of grain reactions. Each grain species takes up one line in this order: reactants, "->", products, "|", key species, Gibb's free energy 'A' term (A/10^4 K), Gibb's free energy 'B' term, surface energy of the condensate (ergs/cm^2), radius of condensate (angstroms). 
 
-*abundance_file*: This lists the names of gas species in the header. Each cell has one line listing: cell ID, number density for each gas species. 
+*abundance_file*: This lists the names of gas species in the header. Each cell has one line listing: cell ID and number density for each gas species. 
 
 *shock_file*: This contains information on a shock. Each cell has one line: cell ID, the time of the shock, the shock temperature, the shock velocity.
 
@@ -132,7 +132,7 @@ If both are set to '1', both destruction and nucleation are calculated.
 *io_restart_n_steps*: Number of cycles until a restart file is updated. 
 
 ### User Specified Shock Parameters
-*pile_up_factor*: This is used to calculate the increase in density when a shock passes through. The density if multiplied by this number. 
+*pile_up_factor*: This is used to calculate the increase in density when a shock passes through. The density is multiplied by this number. 
 
 *shock_velo*: This is the velocity of the shock that will be applied to all cells. 
 
@@ -146,19 +146,19 @@ If both are set to '1', both destruction and nucleation are calculated.
 # Functionality
[…]
