### 2025-07-15 #8349 @guadabsb15
Hi, thanks for the updates. Here are my initial comments about the manuscript after re-checking the most updated version:
1. There are still missing references, I see question marks instead of references in lines 54, 57, and 66.
2. Is the name of the other code `sndust` or `snDust`? It would be nice to keep the naming (capitalization) consistent.
3. You mention the CNT and KNT methods, but it is not clear to me which method is used in nuDustC++
