### 2025-07-15 f5d901c
Update paper.bib
hopefully fixed reference issues
