+++ paper.bib
@@ -30,7 +30,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{kozasa1987grain,
+@article{kozasa1987,
   title={Grain formation through nucleation process in astrophysical environments. II: Nucleation and grain growth accompanied by chemical reaction},
   author={Kozasa, Takahashi and Hasegawa, Hiroichi},
   journal={Progress of Theoretical Physics},
@@ -78,7 +78,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2023ApJ...957...35P,
+@ARTICLE{paquette2023,
        author = {{Paquette}, John A. and {Nuth}, Joseph A. and {Ferguson}, Frank T.},
         title = "{Dust Coagulation in Oxygen-rich Circumstellar Outflows}",
       journal = {The Astrophysical Journal},
@@ -110,7 +110,7 @@ archivePrefix = {arXiv},
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%
 % Nozawqa 2006 earlier shock destruction paper
-@article{Nozawa_2006,
+@article{Nozawa2006,
 	doi = {10.1086/505639},
 	url = {https://doi.org/10.1086%2F505639},
 	year = 2006,
