### 2025-07-15 c8fe252
Update paper.md
fixed snDust vs sndust issue. naming scheme issue
