### 2025-07-15 43bbafe
Update paper.md
specified KNT
