### 2025-08-12 6b35891
Update paper.md
described solution length
