### 2025-08-12 3aa54a9
Update README.md
fixed capitalization, grammar issues
