### 2025-08-12 b12baa2
Update README.md
added more on inputs
