### 2025-09-02 e8cca60
Edited paper to remove performance section. removed the plog take in the CMake file
