### 2025-11-05 9b2cbf4
work for scaling benchmark
