+++ data/benchmark/bm_config.ini
@@ -1,0 +1,17 @@
+ode_dt_0 = 1.0E-2
+ode_dt_min = 1.0E-12
+ode_dt_max = 1
+ode_abs_err = 1.0E-8
+ode_rel_err = 1.0E-8
+io_dump_n_steps = 10000000
+io_restart_n_steps = 10000000
+network_file = data/chemistry/test_chm.chm
+abundance_file = data/benchmark/bm_abundances.dat
+environment_file = data/benchmark/bm_traj.dat
+sizeDist_file = data/size_dist/test_sd.dat
+size_dist_min_rad_exponent_cm = -8
+size_dist_max_rad_exponent_cm = -2
+number_of_size_bins = 100
+do_nucleation = 1
+do_destruction = 0
+mod_number = 2
